Implemented 'OffsetPerID' concept for UnityApplication.
This is used as a workaround for not being able to assign an (application) scene per camera in the TestApp.
We just apply a the 'offset per id' (which is just a big spatial offset, based on the application ID), to the content of each application scene.
This results in the contents of the 'other' application scenes being far away (and thus culled) for each application camera.